Revert "Decompose CLI"
This reverts commit 4c76f3d067660422fbda08346a0cc5176b303939.
